Create own testing module for contour_statistics.py.
